// src/types.ts
export interface ProjectDetailProp {
  name: string;
  description: string;
  date: string;
  rocall: string;
}
export interface Transaction {
  name: string;
  status: string;
  date: string;
  amount: number;
}

export interface PersonnelProp {
  img: string;
  name: string;
  position: string;
}